requirements.txt
selenium~=4.1.3
requests~=2.27.1
behave~=1.2.6
webdriver-manager==3.5.4
PyHamcrest==2.0.3
jsonschema~=4.4.0
jsonpath_rw
